% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_opened_shops_count.R
\name{graph_opened_shops_count}
\alias{graph_opened_shops_count}
\title{Opened Shops}
\usage{
graph_opened_shops_count(year)
}
\description{
A function which, for a given year, gathers information on all shops opened within each month that year.
}
\author{
Lukasz Marchel
}
